Add support for multiple CML hosts.
When using HTTP transport, allow for the client to indicate the CML host
to which they want to connect.  This can be dangerous, so enforce that
an allowed set of hosts is configured if the custom host header is seen.

This was incorporated from parts of #8, but the default to allow all
hosts was changed.